+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./base64.js";
 export * from "./fake-hash.js";
+export * from "./hash-parts.js";